+++ R/clonalCluster.R
@@ -6,6 +6,8 @@
 #' \code{\link{combineBCR}} or \code{\link{combineExpression}} and amends a 
 #' cluster to the data frame or meta data. If \strong{exportGraph} is set 
 #' to TRUE, the function returns an igraph object of the connected sequences. 
+#' If multiple sequences per chain are present, this function only compares
+#' the first sequence.
 #' 
 #' @examples
 #' # Getting the combined contigs
@@ -33,13 +35,15 @@
 #' sequences (\strong{TRUE}) or the amended input with a
 #' new cluster-based variable (\strong{FALSE}).
 #' @importFrom stringdist stringdist
-#' @importFrom igraph set_vertex_attr V
+#' @importFrom igraph set_vertex_attr V union
 #' @importFrom plyr join
-#' @importFrom dplyr bind_rows
+#' @importFrom dplyr bind_rows summarize
 #' @importFrom stringr str_split str_replace_all
 #' @importFrom rlang %||%
 #' @importFrom SummarizedExperiment colData<- colData
 #' @importFrom stats na.omit
+#' @importFrom S4Vectors DataFrame
+#' 
 #' @export
 #' @concept Visualizing_Clones
 #' @return Either amended input with edit-distanced clusters added 
@@ -95,13 +99,19 @@ clonalCluster <- function(input.data,
   
   if (!is.null(group.by)) {
     bound <- bind_rows(dat, .id = "group.by")
+    bound <- bound[!is.na(bound[,ref2]),]
+    retain.ref <- data.frame(old = bound[,ref2], new = str_split(bound[,ref2], ";", simplify = TRUE)[,1])
+    bound[,ref2] <- str_split(bound[,ref2], ";", simplify = TRUE)[,1]
     graph.variables <- bound %>%
                           group_by(bound[,ref2]) %>%
                           dplyr::summarize(sample_count = n(),
                                     unique_samples = paste0(unique(group.by), collapse = ","))
   } else {
     bound <- bind_rows(dat)
-    graph.variables <- bind_rows(dat) %>%
+    bound <- bound[!is.na(bound[,ref2]),]
+    retain.ref <- data.frame(old = bound[,ref2], new = str_split(bound[,ref2], ";", simplify = TRUE)[,1])
+    bound[,ref2] <- str_split(bound[,ref2], ";", simplify = TRUE)[,1]
+    graph.variables <- bound %>%
                           group_by(bound[,ref2]) %>%
                           dplyr::summarize(sample_count = n())
   }
@@ -118,12 +128,12 @@ clonalCluster <- function(input.data,
   #Grabbing column order for later return
   column.order <- colnames(bound)
   
-  #Returning the igraph object if eexportGraph = TRUE
+  #Returning the igraph object if exportGraph = TRUE
   if(exportGraph) {
-    cluster <- output.list[[1]]
+    cluster <- do.call(igraph::union, output.list)
     vertex <- names(V(cluster))
     data_df <- unique(data.frame(
-      id = V(cluster)$name
+      id = vertex
     ))
     data_df <- merge(data_df, graph.variables, by = 1)
     cluster <- set_vertex_attr(cluster, 
@@ -169,10 +179,11 @@ clonalCluster <- function(input.data,
       col.name <- names(PreMeta) %||% colnames(PreMeta)
       input.data[[col.name]] <- PreMeta
     } else {
-      rownames <- rownames(colData(input.data))
-      colData(input.data) <- cbind(colData(input.data), 
-                                   PreMeta[rownames,])[, union(colnames(colData(input.data)),  colnames(PreMeta))]
-      rownames(colData(input.data)) <- rownames 
+      combined_col_names <- unique(c(colnames(colData(sc.data)), colnames(PreMeta)))
+      full_data <- merge(colData(sc.data), PreMeta[rownames, , drop = FALSE], by = "row.names", all.x = TRUE)
+      rownames(full_data) <- full_data[, 1]
+      full_data  <- full_data[, -1]
+      colData(sc.data) <- DataFrame(full_data[, combined_col_names])
     }
   } else {
     #Reorder columns