+++ R/combineContigs.R
@@ -43,8 +43,10 @@ utils::globalVariables(c(
 #' @param removeMulti This will remove barcodes with greater than 2 chains.
 #' @param filterMulti This option will allow for the selection of the 2 
 #' corresponding chains with the highest expression for a single barcode. 
+#' @param filterNonproductive This option will allow for the removal of 
+#' nonproductive chains if the variable exists in the contig data. Default
+#' is set to TRUE to remove nonproductive contigs.
 #' 
-#' @import dplyr
 #' @export
 #' @concept Loading_and_Processing_Contigs
 #' @return List of clones for individual cell barcodes
@@ -54,7 +56,8 @@ combineTCR <- function(input.data,
                        ID = NULL, 
                        removeNA = FALSE, 
                        removeMulti = FALSE, 
-                       filterMulti = FALSE) {
+                       filterMulti = FALSE,
+                       filterNonproductive = TRUE) {
     input.data <- .checkList(input.data)
     input.data <- .checkContigs(input.data)
     out <- NULL
@@ -63,7 +66,7 @@ combineTCR <- function(input.data,
         if(c("chain") %in% colnames(input.data[[i]])) {
           input.data[[i]] <- subset(input.data[[i]], chain != "Multi")
         }
-        if(c("productive") %in% colnames(input.data[[i]])) {
+        if(c("productive") %in% colnames(input.data[[i]]) & filterNonproductive) {
           input.data[[i]] <- subset(input.data[[i]], productive %in% c(TRUE, "TRUE", "True", "true"))
         }
         input.data[[i]]$sample <- samples[i]
@@ -95,13 +98,22 @@ combineTCR <- function(input.data,
         Con.df[Con.df == "NA_NA" | Con.df == "NA;NA_NA;NA"] <- NA 
         data3 <- merge(data2[,-which(names(data2) %in% c("TCR1","TCR2"))], 
             Con.df, by = "barcode")
-        if (!is.null(samples) && !is.null(ID)) {
-            data3 <- data3[, c("barcode", "sample", "ID", tcr1_lines, tcr2_lines,
-                CT_lines)] }
-        else if (!is.null(samples) & is.null(ID)) {
-          data3<-data3[,c("barcode","sample",tcr1_lines,tcr2_lines,
-                          CT_lines)] 
+        
+        columns_to_include <- c("barcode")
+        # Conditionally add columns based on user input
+        if (!is.null(samples)) {
+          columns_to_include <- c(columns_to_include, "sample")
         }
+        if (!is.null(ID)) {
+          columns_to_include <- c(columns_to_include, "ID")
+        }
+        
+        # Add TCR and CT lines which are presumably always needed
+        columns_to_include <- c(columns_to_include, tcr1_lines, tcr2_lines, CT_lines)
+        
+        # Subset the data frame based on the dynamically built list of columns
+        data3 <- data3[, columns_to_include]
+        
         final[[i]] <- data3 
     }
     name_vector <- character(length(samples))
@@ -158,7 +170,7 @@ combineTCR <- function(input.data,
 #' 
 #' @param input.data List of filtered contig annotations or outputs from 
 #' \code{\link{loadContigs}}.
-#' @param samples The labels of samples
+#' @param samples The labels of samples (required).
 #' @param ID The additional sample labeling (optional).
 #' @param call.related.clones Use the nucleotide sequence and V gene 
 #' to call related clones. Default is set to TRUE. FALSE will return 
@@ -169,7 +181,10 @@ combineTCR <- function(input.data,
 #' @param removeMulti This will remove barcodes with greater than 2 chains.
 #' @param filterMulti This option will allow for the selection of the 
 #' highest-expressing light and heavy chains, if not calling related clones.
-#' @import dplyr
+#' @param filterNonproductive This option will allow for the removal of 
+#' nonproductive chains if the variable exists in the contig data. Default
+#' is set to TRUE to remove nonproductive contigs.
+#' @importFrom dplyr %>% mutate
 #' @export
 #' @concept Loading_and_Processing_Contigs
 #' @return List of clones for individual cell barcodes
@@ -180,7 +195,11 @@ combineBCR <- function(input.data,
                        threshold = 0.85,
                        removeNA = FALSE, 
                        removeMulti = FALSE,
-                       filterMulti = TRUE) {
+                       filterMulti = TRUE,
+                       filterNonproductive = TRUE) {
+    if(is.null(samples)) {
+      stop("combineBCR() requires the samples paramter for the calculation of edit distance.")
+    }
     input.data <- .checkList(input.data)
     input.data <- .checkContigs(input.data)
     out <- NULL
@@ -190,6 +209,9 @@ combineBCR <- function(input.data,
     for (i in seq_along(input.data)) {
         input.data[[i]] <- subset(input.data[[i]], chain %in% c("IGH", "IGK", "IGL"))
         input.data[[i]]$ID <- ID[i]
+        if(c("productive") %in% colnames(input.data[[i]]) & filterNonproductive) {
+          input.data[[i]] <- subset(input.data[[i]], productive %in% c(TRUE, "TRUE", "True", "true"))
+        }
         if (filterMulti) {
                     # Keep IGH / IGK / IGL info in save_chain
                     input.data[[i]]$save_chain <- input.data[[i]]$chain